feat: merge action types and utils
