+++ account-kit/smart-contracts/src/index.ts
@@ -118,15 +118,16 @@ export {
 // 6900 RI exports
 export type * from "./6900-v08-RI/account.js";
 export { accountLoupeV08Actions } from "./6900-v08-RI/actions/account-loupe/decorator.js";
-export type * from "./6900-v08-RI/actions/account-loupe/types.js";
+export type * from "./6900-v08-RI/actions/common/types.js";
+export {
+  serializeValidationConfig,
+  deserializeValidationConfig,
+  serializeModuleEntity,
+  deserializeModuleEntity,
+  serializeHookConfig,
+  deserializeHookConfig,
+} from "./6900-v08-RI/actions/common/utils.js";
 export { installValidationActions } from "./6900-v08-RI/actions/install-validation/decorator.js";
-// todo: rectify duplicate types
-export type {
-  ModuleEntity as ModuleEntityT,
-  ValidationConfig as ValidationConfigT,
-  HookType,
-  HookConfig as HookConfigT,
-} from "./6900-v08-RI/actions/install-validation/types.js";
 export { createSingleSignerRIAccount } from "./6900-v08-RI/account.js";
 export { getDefaultSingleSignerRIAccountFactoryAddress } from "./6900-v08-RI/utils.js";
 export { SingleSignerValidationModule } from "./6900-v08-RI/modules/single-signer-validation/module.js";
