+++ site/pages/reference/account-kit/smart-contracts/functions/deserializeHookConfig.mdx
@@ -1,0 +1,36 @@
+---
+# This file is autogenerated
+title: deserializeHookConfig
+description: Overview of the deserializeHookConfig method
+---
+
+# deserializeHookConfig
+
+Deserializes a given hex string into a HookConfig object.
+The hex string must be a 26-byte hex string, prefixed with `0x`.
+
+## Import
+
+```ts
+import { deserializeHookConfig } from "@account-kit/smart-contracts";
+```
+
+## Usage
+
+```ts
+const config = deserializeHookConfig(
+  "0xAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAA000004D20001"
+);
+```
+
+## Parameters
+
+### hex
+
+`Hex`
+A 26-byte hex string representing the hook configuration, prefixed with `0x`
+
+## Returns
+
+`HookConfig`
+An object representing the deserialized hook configuration
