+++ site/pages/reference/account-kit/smart-contracts/functions/deserializeModuleEntity.mdx
@@ -1,0 +1,35 @@
+---
+# This file is autogenerated
+title: deserializeModuleEntity
+description: Overview of the deserializeModuleEntity method
+---
+
+# deserializeModuleEntity
+
+Deserializes a hexadecimal string into a `ModuleEntity` object containing a module address and an entity ID.
+
+## Import
+
+```ts
+import { deserializeModuleEntity } from "@account-kit/smart-contracts";
+```
+
+## Usage
+
+```ts
+const entity = deserializeModuleEntity(
+  "0xAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAA000004D2"
+);
+```
+
+## Parameters
+
+### hex
+
+`Hex`
+The 24-byte hexadecimal string, prefixed with `0x`, representing the module entity
+
+## Returns
+
+`ModuleEntity`
+An object containing the module address and entity ID
