+++ site/pages/reference/account-kit/smart-contracts/functions/deserializeValidationConfig.mdx
@@ -1,0 +1,35 @@
+---
+# This file is autogenerated
+title: deserializeValidationConfig
+description: Overview of the deserializeValidationConfig method
+---
+
+# deserializeValidationConfig
+
+Deserializes a validation configuration hex string into a `ValidationConfig` object.
+
+## Import
+
+```ts
+import { deserializeValidationConfig } from "@account-kit/smart-contracts";
+```
+
+## Usage
+
+```ts
+const config = deserializeValidationConfig(
+  "0xAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAA000004D201"
+);
+```
+
+## Parameters
+
+### hex
+
+`Hex`
+The 26-byte hex string representation of the validation configuration
+
+## Returns
+
+`ValidationConfig`
+An object containing the module address, entity ID, global status, and signature validation status
