+++ site/pages/reference/account-kit/smart-contracts/functions/serializeHookConfig.mdx
@@ -1,0 +1,40 @@
+---
+# This file is autogenerated
+
+title: serializeHookConfig
+description: Overview of the serializeHookConfig method
+---
+
+# serializeHookConfig
+
+Serializes a `HookConfig` object into a 26-byte hex string.
+
+## Import
+
+```ts
+import { serializeHookConfig } from "@account-kit/smart-contracts";
+```
+
+## Usage
+
+```ts
+const config = serializeHookConfig({
+  moduleAddress: "0xAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAA",
+  entityId: 1234,
+  hookType: HookType.EXECUTION,
+  hasPreHooks: true,
+  hasPostHooks: false,
+});
+```
+
+## Parameters
+
+### config
+
+`HookConfig`
+The hook configuration to serialize
+
+## Returns
+
+`Hex`
+A 26-byte hex string representation of the hook configuration
