+++ site/pages/reference/account-kit/smart-contracts/functions/serializeModuleEntity.mdx
@@ -1,0 +1,37 @@
+---
+# This file is autogenerated
+
+title: serializeModuleEntity
+description: Overview of the serializeModuleEntity method
+---
+
+# serializeModuleEntity
+
+Serializes a `ModuleEntity` object into a 24-byte hex string.
+
+## Import
+
+```ts
+import { serializeModuleEntity } from "@account-kit/smart-contracts";
+```
+
+## Usage
+
+```ts
+const entity = serializeModuleEntity({
+  moduleAddress: "0xAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAA",
+  entityId: 1234,
+});
+```
+
+## Parameters
+
+### moduleEntity
+
+`ModuleEntity`
+The module entity to serialize
+
+## Returns
+
+`Hex`
+A 24-byte hex string representation of the module entity
