+++ site/pages/reference/account-kit/smart-contracts/functions/serializeValidationConfig.mdx
@@ -1,0 +1,39 @@
+---
+# This file is autogenerated
+
+title: serializeValidationConfig
+description: Overview of the serializeValidationConfig method
+---
+
+# serializeValidationConfig
+
+Serializes a `ValidationConfig` object into a 26-byte hex string.
+
+## Import
+
+```ts
+import { serializeValidationConfig } from "@account-kit/smart-contracts";
+```
+
+## Usage
+
+```ts
+const config = serializeValidationConfig({
+  moduleAddress: "0xAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAA",
+  entityId: 1234,
+  isGlobal: true,
+  isSignatureValidation: true,
+});
+```
+
+## Parameters
+
+### config
+
+`ValidationConfig`
+The validation configuration to serialize
+
+## Returns
+
+`Hex`
+A 26-byte hex string representation of the validation configuration
